+++ lib/src/animation/toggle_rotate.dart
@@ -1,10 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_waya/src/extended_state.dart';
 
-typedef ToggleRotateFunction = void Function({bool reset});
+typedef ToggleRotateHandleRotation = void Function(
+    {bool reset, bool needsBuild});
 
 typedef ToggleRotateBuilder = Widget Function(
-    Widget child, ToggleRotateFunction rotate);
+    Widget child, ToggleRotateHandleRotation rotate);
+
+typedef ToggleRotateIconBuilder = Widget Function(bool isRotation);
+
+typedef ToggleRotateCallback = void Function(bool isRotation);
+
+extension ExtensionWidgetToggleRotate on Widget {
+  ToggleRotateIconBuilder get toToggleRotateIconBuilder => (isRotation) => this;
+}
 
 /// 旋转组件
 class ToggleRotate extends StatefulWidget {
@@ -16,10 +25,17 @@ class ToggleRotate extends StatefulWidget {
       this.clockwise = true,
       this.duration = const Duration(milliseconds: 200),
       this.curve = Curves.fastOutSlowIn,
-      this.isRotate = false});
+      this.isRotate = false,
+      this.onChanged});
 
   /// 旋转的icon
-  final Widget icon;
+  final ToggleRotateIconBuilder icon;
+
+  /// 自定义非旋转区域
+  final ToggleRotateBuilder? builder;
+
+  /// 变化监听
+  final ToggleRotateCallback? onChanged;
 
   /// 是否旋转
   final bool isRotate;
@@ -37,9 +53,6 @@ class ToggleRotate extends StatefulWidget {
   /// 动画曲线
   final Curve curve;
 
-  /// 自定义非旋转区域
-  final ToggleRotateBuilder? builder;
-
   @override
   State<ToggleRotate> createState() => _ToggleRotateState();
 }
@@ -52,11 +65,20 @@ class _ToggleRotateState extends ExtendedState<ToggleRotate>
   @override
   void initState() {
     controller = AnimationController(duration: widget.duration, vsync: this);
+    controller.addStatusListener(statusListener);
     initTween();
     super.initState();
     WidgetsBinding.instance.addPostFrameCallback((_) {
       if (widget.isRotate) controller.forward();
     });
+  }
+
+  void statusListener(AnimationStatus status) {
+    if (status == AnimationStatus.completed) {
+      widget.onChanged?.call(true);
+    } else if (status == AnimationStatus.dismissed) {
+      widget.onChanged?.call(false);
+    }
   }
 
   void initTween() {
@@ -70,19 +92,15 @@ class _ToggleRotateState extends ExtendedState<ToggleRotate>
         .chain(CurveTween(curve: widget.curve)));
   }
 
-  @override
-  void dispose() {
-    controller.dispose();
-    super.dispose();
-  }
-
-  void handleRotation({bool reset = false}) {
+  void handleRotation({bool reset = false, bool needsBuild = false}) {
     if (reset) controller.reset();
     if (controller.value == 0.0) {
       controller.forward();
     } else {
       controller.reverse();
     }
+
+    if (needsBuild) setState(() {});
   }
 
   @override
@@ -99,10 +117,16 @@ class _ToggleRotateState extends ExtendedState<ToggleRotate>
   Widget build(BuildContext context) {
     final current = AnimatedBuilder(
         animation: controller.view,
-        child: widget.icon,
-        builder: (_, Widget? child) =>
-            RotationTransition(turns: rotate, child: child));
+        builder: (_, Widget? child) => RotationTransition(
+            turns: rotate, child: widget.icon(controller.value == 1.0)));
     if (widget.builder == null) return current;
     return widget.builder!(current, handleRotation);
+  }
+
+  @override
+  void dispose() {
+    controller.removeStatusListener(statusListener);
+    controller.dispose();
+    super.dispose();
   }
 }